pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clawmeets"
version = "1.1.7"
description = "Agent runner for clawmeets multi-agent collaboration"
license = "MIT"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "httpx>=0.27",
    "websockets>=12",
    "typer>=0.12",
    "pydantic>=2.7",
    "aiofiles>=23",
    "bcrypt>=4.0.0",
    # MCP SDK + Google client stack are part of the baseline install so the
    # in-tree Gmail and Google Calendar MCP servers, and `clawmeets mcp auth`,
    # work out of the box. Collectively ~15 MB on disk — small enough that
    # gating them behind optional extras costs more user friction than it saves.
    "mcp>=1.0",
    "google-api-python-client>=2.100",
    "google-auth>=2.23",
    "google-auth-oauthlib>=1.1",
    "python-chess>=1.10",
]

[project.scripts]
clawmeets = "clawmeets.cli:main"

[project.urls]
Homepage = "https://github.com/clawmeets-ai/clawmeets"
Repository = "https://github.com/clawmeets-ai/clawmeets"
Issues = "https://github.com/clawmeets-ai/clawmeets/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["clawmeets*"]

[tool.setuptools.package-data]
"clawmeets.mcp.servers" = ["*.html"]
